bittorrent_deployer/requirements.txt
# AWS Deployment Dependencies
boto3>=1.26.0
botocore>=1.29.0

# Configuration Management
pyyaml>=6.0

# HTTP Requests
requests>=2.28.0

# Concurrent Processing
# (concurrent.futures is built-in since Python 3.2)